clean up auth routes
